Anmol

+++ src/main/java/com/ats/adminpanel/controller/BillController.java
@@ -337,7 +337,7 @@ public class BillController {
 						Float baseRate = (orderRate * 100) / (100 + (tax1 + tax2));
 						baseRate = roundUp(baseRate);
 
-						Float taxableAmt = (float) (baseRate * Integer.parseInt(billQty));
+						Float taxableAmt = (float) (baseRate * Float.parseFloat(billQty));
 
 						System.out.println("taxableAmt: " + taxableAmt);
 						taxableAmt = roundUp(taxableAmt);
@@ -353,7 +353,7 @@ public class BillController {
 
 						if (gBill.getIsSameState() == 1) {
 							baseRate = (orderRate * 100) / (100 + (tax1 + tax2));
-							taxableAmt = (float) (baseRate * Integer.parseInt(billQty));
+							taxableAmt = (float) (baseRate * Float.parseFloat(billQty));
 							// ----------------------------------------------------------
 							discAmt = ((taxableAmt * discPer) / 100); // new row added
 							System.out.println("discAmt: " + discAmt);// new row added
@@ -370,7 +370,7 @@ public class BillController {
 
 						else {
 							baseRate = (orderRate * 100) / (100 + (tax3));
-							taxableAmt = (float) (baseRate * Integer.parseInt(billQty));
+							taxableAmt = (float) (baseRate * Float.parseFloat(billQty));
 							// ----------------------------------------------------------
 							discAmt = ((taxableAmt * discPer) / 100); // new row added
 							System.out.println("discAmt: " + discAmt);// new row added
@@ -411,7 +411,7 @@ public class BillController {
 						billDetail.setCatId(gBill.getCatId());
 						billDetail.setItemId(gBill.getItemId());
 						billDetail.setOrderQty(gBill.getOrderQty());
-						billDetail.setBillQty(Integer.parseInt(billQty));
+						billDetail.setBillQty(Float.parseFloat(billQty));
 						billDetail.setMrp((float) gBill.getOrderMrp());
 						billDetail.setRateType(gBill.getRateType());
 						billDetail.setRate((float) gBill.getOrderRate());
